cleaned up padding and spacing of scene items for better standards, removed notification dropdown and opted for standalone buttons for both subscribe and alerts
