+++ common/tailwind/tailwind.css
@@ -6,4 +6,6 @@
 @theme {
     --text-xxs: 0.625rem;
     --text-xxs--line-height: 0.75rem;
+
+    --spacing-button-padding-x: 6px; /* Match the padding of the button primitives base size */
 }
