+++ frontend/src/layout/scenes/SceneLayout.tsx
@@ -42,26 +42,29 @@ export function ScenePanelDivider(): JSX.Element {
     return <LemonDivider className="-mx-2 my-2 w-[calc(100%+1rem)]" />
 }
 
-export function ScenePanelMetaInfo({ children }: { children: React.ReactNode }): JSX.Element {
-    return <div className="pl-1 pt-4 pb-2 flex flex-col gap-2">{children}</div>
-}
-
+// Should be first!
 export function ScenePanelCommonActions({ children }: { children: React.ReactNode }): JSX.Element {
     return (
         <>
-            <div className="flex flex-col gap-2">{children}</div>
+            <div className="flex flex-col gap-2 pt-2">{children}</div>
             <ScenePanelDivider />
         </>
     )
 }
 
+// Should be second!
+export function ScenePanelMetaInfo({ children }: { children: React.ReactNode }): JSX.Element {
+    return <div className="pl-1 pb-1 flex flex-col gap-2 -ml-1">{children}</div>
+}
+
+// Should be third!
 export function ScenePanelActions({ children }: { children: React.ReactNode }): JSX.Element {
     return (
-        <div className="flex flex-col gap-2">
-            <Label intent="menu" className="px-1">
+        <div className="flex flex-col gap-2 pl-1 -ml-1">
+            <Label intent="menu" className="mx-2">
                 Actions
             </Label>
-            <div className="flex flex-col gap-px">{children}</div>
+            <div className="flex flex-col gap-px -ml-1 pl-1">{children}</div>
         </div>
     )
 }
@@ -70,7 +73,7 @@ export function ScenePanelLabel({ children, title, ...props }: PropsWithChildren
     return (
         <div>
             <div className="gap-0">
-                <Label intent="menu" {...props}>
+                <Label intent="menu" {...props} className="mx-button-padding-x">
                     {title}
                 </Label>
                 {children}
