+++ frontend/src/lib/components/Scenes/SceneDescription.tsx
@@ -39,7 +39,7 @@ export function SceneDescription({
     return localIsEditing ? (
         <form onSubmit={handleSubmit} name="page-description-form" className="flex flex-col gap-1">
             <div className="flex flex-col gap-0">
-                <Label intent="menu" htmlFor="page-description-input">
+                <Label intent="menu" htmlFor="page-description-input" className="mx-2">
                     Description
                 </Label>
                 <TextareaPrimitive
@@ -52,7 +52,6 @@ export function SceneDescription({
                     data-attr={`${dataAttrKey}-description-input`}
                     autoFocus
                     error={!!error}
-                    className="-ml-1.5"
                 />
             </div>
             <div className="flex gap-1">
@@ -80,23 +79,21 @@ export function SceneDescription({
         </form>
     ) : (
         <ScenePanelLabel title="Description">
-            <div className="-ml-1.5">
-                <ButtonPrimitive
-                    className="hyphens-auto flex gap-1 items-center"
-                    lang="en"
-                    onClick={() => setLocalIsEditing(true)}
-                    tooltip={canEdit ? 'Edit description' : 'Description is read-only'}
-                    autoHeight
-                    menuItem
-                    inert={!canEdit}
-                >
-                    {defaultValue !== '' ? (
-                        defaultValue
-                    ) : (
-                        <span className="text-tertiary font-normal">No description {optional ? '(optional)' : ''}</span>
-                    )}
-                </ButtonPrimitive>
-            </div>
+            <ButtonPrimitive
+                className="hyphens-auto flex gap-1 items-center"
+                lang="en"
+                onClick={() => setLocalIsEditing(true)}
+                tooltip={canEdit ? 'Edit description' : 'Description is read-only'}
+                autoHeight
+                menuItem
+                inert={!canEdit}
+            >
+                {defaultValue !== '' ? (
+                    defaultValue
+                ) : (
+                    <span className="text-tertiary font-normal">No description {optional ? '(optional)' : ''}</span>
+                )}
+            </ButtonPrimitive>
         </ScenePanelLabel>
     )
 }
