+++ frontend/src/lib/components/Scenes/SceneTags.tsx
@@ -47,7 +47,6 @@ export function SceneTags({ onSave, tags, tagsAvailable, dataAttrKey, canEdit = 
                     placeholder='try "official"'
                     size="xsmall"
                     autoFocus
-                    className="-ml-1.5"
                 />
             </ScenePanelLabel>
             <div className="flex gap-1">
@@ -76,30 +75,28 @@ export function SceneTags({ onSave, tags, tagsAvailable, dataAttrKey, canEdit = 
         </form>
     ) : (
         <ScenePanelLabel title="Tags">
-            <div className="-ml-1.5">
-                <ButtonPrimitive
-                    className="hyphens-auto flex gap-1 items-center"
-                    lang="en"
-                    onClick={() => onSave && canEdit && setLocalIsEditing(true)}
-                    tooltip={canEdit ? 'Edit tags' : 'Tags are read-only'}
-                    autoHeight
-                    menuItem
-                    inert={!canEdit}
-                    data-attr={`${dataAttrKey}-tags-button`}
-                >
-                    {tags && tags.length > 0 ? (
-                        <ObjectTags tags={tags} data-attr={`${dataAttrKey}-tags`} staticOnly />
-                    ) : (
-                        <>
-                            {onSave && canEdit ? (
-                                'Click to add tags'
-                            ) : (
-                                <span className="text-tertiary font-normal">No tags</span>
-                            )}
-                        </>
-                    )}
-                </ButtonPrimitive>
-            </div>
+            <ButtonPrimitive
+                className="hyphens-auto flex gap-1 items-center"
+                lang="en"
+                onClick={() => onSave && canEdit && setLocalIsEditing(true)}
+                tooltip={canEdit ? 'Edit tags' : 'Tags are read-only'}
+                autoHeight
+                menuItem
+                inert={!canEdit}
+                data-attr={`${dataAttrKey}-tags-button`}
+            >
+                {tags && tags.length > 0 ? (
+                    <ObjectTags tags={tags} data-attr={`${dataAttrKey}-tags`} staticOnly />
+                ) : (
+                    <>
+                        {onSave && canEdit ? (
+                            'Click to add tags'
+                        ) : (
+                            <span className="text-tertiary font-normal">No tags</span>
+                        )}
+                    </>
+                )}
+            </ButtonPrimitive>
         </ScenePanelLabel>
     )
 }
